server/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib
bcrypt==4.0.1
pyjwt
pydantic
pydantic-settings
